refactored uploader
